+++ History.md
@@ -1,5 +1,8 @@
 # History
 
+#### 2021-01-19
+* Add ability to add or update Source. Closes #39.
+
 #### 2021-01-18
 * Add ability to add or update Category. Closes #38.